Various FPU & VU changes and some fixes (#56)
Implemented FPU flags and control register
Fixed a typo
Fixed link register write size for JALR
Adjusted the VU int registers to have signed and unsigned variants. Adjusted ops as appropriate.
Added a DISPLAY selector to the GS (temp code until it's implemented properly)
Added code to set the CPU rounding mode to Chop/Zero to match the PS2 (Fixes Torus demo)

+++ src/core/ee/emotioninterpreter.cpp
@@ -319,7 +319,7 @@ void EmotionInterpreter::jal(EmotionEngine &cpu, uint32_t instruction)
     uint32_t PC = cpu.get_PC();
     addr += (PC + 4) & 0xF0000000;
     cpu.jp(addr);
-    cpu.set_gpr<uint32_t>(31, PC + 8);
+    cpu.set_gpr<uint64_t>(31, PC + 8);
 }
 
 void EmotionInterpreter::beq(EmotionEngine &cpu, uint32_t instruction)
